Added mapping file for Salvage study
Added mapping file for Salvage study, New file format prepared for DipSeq Rev and Fwd indexes, different that I usually use, with sample ID only having a number 1- 62, and there is extraneous info for this stage, such as the barcode sequence and the well of the plate it was amplified in. The things that are interesting to compare are the Year, Site (which will be the recipient sites: HH, PS, or WL, and the donor site as : DS), Treat (treatment - thickness of topsoil added treatment group), and TSDel (which may not be of interest but is the aliquot of topsoil delivered).